Move StartupConfig to its own cpp/h
