+++ source/ConfigFiles.h
@@ -15,171 +15,6 @@ static bool operator==(Array<T> a1, Array<T> a2) {
 	return !(a1 != a2);
 }
 
-class ConfigFiles {
-public: 
-	String name;								///< Name for the experiment
-	String experimentConfigFilename;			///< Experiment configuration filename/path
-	String userConfigFilename;					///< User configuration filename/path
-	String userStatusFilename;					///< User status filename/path
-	String keymapConfigFilename;				///< Keymap configuration filename/path
-	String systemConfigFilename;				///< System configuration filename/path
-	String resultsDirPath;						///< Results directory path
-
-	ConfigFiles() {};
-
-	ConfigFiles(String name, String expConfig, String userConfig, String userStatus, String keymapConfig, String systemConfig, String resultsDir) :
-		name(name), experimentConfigFilename(expConfig), userConfigFilename(userConfig), userStatusFilename(userStatus), 
-		keymapConfigFilename(keymapConfig), systemConfigFilename(systemConfig), resultsDirPath(resultsDir) {};
-
-	static ConfigFiles defaults() { return ConfigFiles("default", "experimentconfig.Any", "userconfig.Any", "userstatus.Any", "keymap.Any", "systemconfig.Any", "./results"); }
-
-	ConfigFiles(const Any& any) {
-		AnyTableReader reader(any);
-
-		reader.get("name", name, "Must provide a name for every entry in the experimentList!");
-
-		reader.getIfPresent("experimentConfigFilename", experimentConfigFilename);
-		checkValidAnyFilename("experimentConfigFilename", experimentConfigFilename);
-
-		reader.getIfPresent("userConfigFilename", userConfigFilename);
-		checkValidAnyFilename("userConfigFilename", userConfigFilename);
-
-		reader.getIfPresent("userStatusFilename", userStatusFilename);
-		checkValidAnyFilename("userStatusFilename", userStatusFilename);
-
-		reader.getIfPresent("keymapConfigFilename", keymapConfigFilename);
-		checkValidAnyFilename("keymapConfigFilename", keymapConfigFilename);
-
-		reader.getIfPresent("systemConfigFilename", systemConfigFilename);
-		checkValidAnyFilename("systemConfigFilename", systemConfigFilename);
-
-		reader.getIfPresent("resultsDirPath", resultsDirPath);
-		resultsDirPath = formatDirPath(resultsDirPath);
-	}
-
-	/** Assert that the filename `path` ends in .any and report `errorName` if it doesn't */
-	static void checkValidAnyFilename(const String& errorName, const String& path) {
-		if (path.empty()) return;	// Allow empty paths (will be replaced)
-		alwaysAssertM(toLower(path.substr(path.length() - 4)) == ".any", "Config filenames specified in the startup config must end with \".any\"!, check the " + errorName + "!\n");
-	}
-
-	/** Returns the provided path with trailing slashes added if missing */
-	static String formatDirPath(const String& path) {
-		String fpath = path;
-		if (!path.empty() && path.substr(path.length() - 1) != "/") { fpath = path + "/"; }
-		return fpath;
-	}
-
-	Any toAny(const bool forceAll = false) const {
-		Any a(Any::TABLE);
-		a["name"] = name;
-		a["experimentConfigFilename"] = experimentConfigFilename;
-		a["userConfigFilename"] = userConfigFilename;
-		a["userStatusFilename"] = userStatusFilename;
-		a["keymapConfigFilename"] = keymapConfigFilename;
-		a["systemConfigFilename"] = systemConfigFilename;
-		a["resultsDirPath"] = resultsDirPath;
-		return a;
-	}
-
-	void populateEmptyFieldsWithDefaults(const ConfigFiles& def) {
-		if (experimentConfigFilename.empty()) { experimentConfigFilename = def.experimentConfigFilename;  }
-		if (userConfigFilename.empty()) { userConfigFilename = def.userConfigFilename; }
-		if (userStatusFilename.empty()) { userStatusFilename = def.userStatusFilename; }
-		if (keymapConfigFilename.empty()) { keymapConfigFilename = def.keymapConfigFilename; }
-		if (systemConfigFilename.empty()) { systemConfigFilename = def.systemConfigFilename; }
-		if (resultsDirPath.empty()) { resultsDirPath = def.resultsDirPath; }
-	}
-};
-
-/** Configure how the application should start */
-class StartupConfig {
-public:
-	bool	developerMode = false;								///< Sets whether the app is run in "developer mode" (i.e. w/ extra menus)
-	bool	waypointEditorMode = false;							///< Sets whether the app is run w/ the waypoint editor available
-	bool	fullscreen = true;									///< Whether the app runs in windowed mode
-	Vector2 windowSize = { 1920, 980 };							///< Window size (when not run in fullscreen)
-	
-	ConfigFiles defaultExperiment = ConfigFiles::defaults();	///< Setup default list
-	Array<ConfigFiles> experimentList;							///< List of configs (for various experiments)
-	
-	bool	audioEnable = true;									///< Audio on/off
-
-	static StartupConfig load(const String& filename) {
-		StartupConfig config;
-		if (!FileSystem::exists(filename)) {
-			config.toAny(true).save(filename);		// If the file doesn't exist create it
-		}
-		return Any::fromFile(filename);				// Load back from the Any file
-	}
-
-	StartupConfig() {};
-
-	/** Construct from any here */
-    StartupConfig(const Any& any) {
-		bool foundDefault = false;
-        int settingsVersion = 1;
-        AnyTableReader reader(any);
-        reader.getIfPresent("settingsVersion", settingsVersion);
-
-        switch (settingsVersion) {
-        case 1:
-            reader.getIfPresent("developerMode", developerMode);
-			reader.getIfPresent("waypointEditorMode", waypointEditorMode);
-			reader.getIfPresent("fullscreen", fullscreen);
-			reader.getIfPresent("windowSize", windowSize);
-
-			foundDefault = reader.getIfPresent("defaultExperiment", defaultExperiment);
-			if (!foundDefault) {
-				logPrintf("Warning: no `defaultExperiment` found in `startupConfig.Any`, you may be using an old experiment config.\n");
-				String experimentConfigFilename, userConfigFilename, userStatusFilename, keymapConfigFilename, systemConfigFilename, resultsDirPath;
-				reader.getIfPresent("experimentConfigFilename", experimentConfigFilename);
-				reader.getIfPresent("userConfigFilename", userConfigFilename);
-				reader.getIfPresent("userStatusFilename", userStatusFilename);
-				reader.getIfPresent("keymapConfigFilename", keymapConfigFilename);
-				reader.getIfPresent("systemConfigFilename", systemConfigFilename);
-				reader.getIfPresent("resultsDirPath", resultsDirPath);
-				logPrintf("One possible fix is to use the following in `startupConfig.Any`:\n");
-				logPrintf("    defaultExperiment = {\n");
-				logPrintf("        name = \"default\";\n");
-				logPrintf("        experimentConfigFilename = \"%s\";\n", experimentConfigFilename);
-				logPrintf("        userConfigFilename = \"%s\";\n", userConfigFilename);
-				logPrintf("        userStatusFilename = \"%s\";\n", userStatusFilename);
-				logPrintf("        keymapConfigFilename = \"%s\";\n", keymapConfigFilename);
-				logPrintf("        systemConfigFilename = \"%s\";\n", systemConfigFilename);
-				logPrintf("        resultsDirPath = \"%s\";\n", resultsDirPath);
-				logPrintf("    };\n");
-			}
-
-			reader.getIfPresent("experimentList", experimentList);
-			for (ConfigFiles& files : experimentList) { files.populateEmptyFieldsWithDefaults(defaultExperiment); }
-			if (experimentList.length() == 0) { experimentList.append(defaultExperiment); }
-
-            reader.getIfPresent("audioEnable", audioEnable);
-            break;
-        default:
-            debugPrintf("Settings version '%d' not recognized in StartupConfig.\n", settingsVersion);
-            break;
-        }
-    }
-
-	/** Allow this to be converted back to any */
-    Any toAny(const bool forceAll = true) const {
-		StartupConfig def;		// Create a dummy default config for value testing
-        Any a(Any::TABLE);
-        if(forceAll || def.developerMode != developerMode)								a["developerMode"] = developerMode;
-		if(forceAll || def.waypointEditorMode != waypointEditorMode)					a["waypointEditorMode"] = waypointEditorMode;
-		if(forceAll || def.fullscreen != fullscreen)									a["fullscreen"] = fullscreen;
-        if(forceAll || def.audioEnable != audioEnable)									a["audioEnable"] = audioEnable;
-		a["defaultExperiment"] = defaultExperiment;
-		a["experimentList"] = experimentList;
-		
-		return a;
-    }
-
-
-};
-
 /** Key mapping */
 class KeyMapping {
 public:
