+++ vs/FPSci.lib.vcxproj
@@ -152,6 +152,7 @@
     <ClInclude Include="..\source\PlayerEntity.h" />
     <ClInclude Include="..\source\PythonLogger.h" />
     <ClInclude Include="..\source\sqlHelpers.h" />
+    <ClInclude Include="..\source\StartupConfig.h" />
     <ClInclude Include="..\source\TargetEntity.h" />
     <ClInclude Include="..\source\GuiElements.h" />
     <ClInclude Include="..\source\Dialogs.h" />
@@ -165,6 +166,7 @@
     <ClCompile Include="..\source\PhysicsScene.cpp" />
     <ClCompile Include="..\source\PlayerEntity.cpp" />
     <ClCompile Include="..\source\sqlHelpers.cpp" />
+    <ClCompile Include="..\source\StartupConfig.cpp" />
     <ClCompile Include="..\source\TargetEntity.cpp" />
     <ClCompile Include="..\source\GuiElements.cpp" />
     <ClCompile Include="..\source\WaypointManager.cpp" />
